Fix pip-compile args: force upgrade to latest versions
I noticed that unless you pass the `-U` ("Try to upgrade all dependencies to
their latest versions") flag to `pip-compile`, that it will not update any
packages if there is a corresponding `.txt` file to the `.in` file you call it
on as long as that `.txt` file containing all it's dependencies. Further, even
if it contains now-deprecated dependencies, those will not be removed from this
`.txt` file. I think this behavior is on purpose to mimic pip in how pip will
not upgrade any packages existing packages unless `-U` is passed. Anyway, this
resolves the issue.