Fix pip-compile args: force upgrade to latest versions
I noticed that unless you pass the `-U` ("Try to upgrade all dependencies to
their latest versions") flag to `pip-compile`, that it will not update any
packages if there is a corresponding `.txt` file to the `.in` file you call it
on as long as that `.txt` file containing all it's dependencies. Further, even
if it contains now-deprecated dependencies, those will not be removed from this
`.txt` file. I think this behavior has something to do with the logic that
allows one to compile two `.in` files into a single `.txt` file, however, that's
just a hunch. Anyway, this resolves the issue.