Comment out apparmor lines in sd-app-code build and postinst
Apparmor looks for /var/log/syslog when aa-enforce is run in the postinst
script. This triggers an exception in the postinst script because this file does
not exist in our containers. For now, we'll just remove the apparmour code to
get a working container image. Later, it will be re-added.