Updates IPv4 address inference in staging vars
In production installs of SecureDrop, the local IPv4 addresses must be
hard-coded by the Admins in the vars files. These addresses are used
to connect to the hosts during the initial Ansible run, after which SSH
access is restricted to Authenticated Tor Hidden Service (ATHS) for SSH.
The IPv4 addresses are also used for OSSEC monitoring over the LAN,
behind the hardware firewall.

In the staging environment, we have the benefit of inferring the
internal IPv4 addresses in testing environments, since developers
control the machine creation. Although it's possible to hardcode IP
addresses for the Vagrant VMs, and use these hardcoded values in the
staging vars, the same approach would not work for testing in
environments other than Vagrant/VirtualBox, such as DigitalOcean.

The previous approach naively used the Ansible host fact
`.ansible_all_ipv4_addresses.0`, which is *usually* correct, but not
guaranteed to be, since the list var of all IPv4 addresses is not
ordered. Therefore let's specify the interface we expect the IPv4
address to be associated with, and explicitly retrieve the address based
on the interface name. This allows for selective overrides in different
environments, e.g. VirtualBox vs. DigitalOcean. Production vars are
unaffected by this change.