Creates securedrop-keyring deb package directory structure
Simple as it gets: store a binary keyring export in the directory
structure of apt's keyrings. The existing build package roles in the
securedrop repo handle bottling up the files in a deb package just fine.

The keyring file was generated using:

  gpg --no-default-keyring --keyring \
  install_files/securedrop-keyring/etc/apt/trusted.gpg.d/securedrop-keyring.gpg \
  --import \
  install_files/ansible-base/roles/install_fpf_repo/files/fpf-signing-key.pub

The keyring changes constitute a transition between fingerprints,
specifically:

  FROM  B89A29DB2128160B8E4B1B4CBADDE0C7FC9F6818
  TO    22245C81E3BAEB4138B36061310F561200F4AD77

Noting these changes here because the keyring file is binary format, and
therefore difficult to inspect changes for.