Fix validation for editing usernames
Previously, if an admin tried to edit a user and give them a username
that was already in use, the page would flash a highly unhelpful
"unknown error occurred" notification. This validation was also being
triggered indirectly from an exception raised by SQLAlchemy.

This commit makes username validation explicit for editing users from
the admin interface and flashes a specific and helpful message.

Updates the unit tests to check for the new message.

Also fixes an issue where a "password successfully changed" message was
flashed any time the user was edited. Fixed so this is only shown when
the password is actually successfully changed.