Updates Serverspec tests for FPF/SD GPG apt signing key
The changes in 0.3.10 rotate the signing key used for the
apt.freedom.press repo, from which SecureDrop-related deb packages are
installed. Updated the Serverspec tests to confirm explicit presence of
the new signing key, in a specific file, and to confirm absence of the
old fingerprint, including a search just for the old fingerprint.