Developer docs: updates base package installation command
Updating supported Ubuntu OS to 16.04, which is the latest LTS and also
the only version of Ubuntu that's recently been tested against these
docs. In both Debian Stretch and Ubuntu Xenial (16.04), the package
`virtualbox` depends on `virtualbox-dkms`, so listing the former as the
install target is more accurate.

Also removes manual invocation of `dpkg-reconfigure virtualbox-dkms`,
since the postinst hooks will set up the vbox driver automatically.