Developer docs: makes vagrant version warnings terser
Let's inline the relevant GitHub issues showing problems with
out-of-date versions of Vagrant, and keep going with the install docs.
Being too repetitious about the important of the recommended version
hurts readability of the docs.

Also explicit mentions a specific version of 5.x for VirtualBox, since
we know that the 4.x release series will not work.