Names VirtualBox internal network from directory name
Since the IPv4 addresses for the staging and prod hosts are hardcoded in
the Vagrantfile, multiple instances of the VMs will try to reuse the IP
on the same internal network. By default, the internal network that
VirtualBox creates is called "intnet", so the IP reuse will cause
failure for the App and Mon servers communicating with each other for
OSSEC functionality.

The most likely use case for multiple concurrent VM instances is in
conjunction with `git worktree`, which will create a new working copy of
an existing local git repository. It's a bit of an assumption that the
two worktrees will have a unique basenames (e.g. "securedrop" and
"securedrop2"), but a reasonable one—more reasonable than using the
absolute path to the repo, which is certain to be unique, but extremely
verbose.

More info on VirtualBox internal networks here:

  * https://www.virtualbox.org/manual/ch06.html#network_internal
  * https://www.vagrantup.com/docs/virtualbox/networking.html