Updates AppArmor profile to permit reading new font-awesome icons
The SVG -> PNG conversion in a3e791084b92d4362dec12b59b104b4cf56bbc2d
requires that the explicit filepaths in the AppArmor profile for apache2
be updated. The SVG paths were never present (because the changes only
landed in develop, and never shipped as part of a release), so this is
an additive diff, rather than an edit.

This AppArmor profile will be included in the securedrop-app-code deb
package, and shipped with the 0.3.11 release.