Remove instructions for using Tor Browser since sources are already using Tor Browser
This link is no longer needed. @garrettr noted that this link was
originally here to link to the page that contained instructions
on disabling JS in other browsers than Tor Browser, but those
days are over since if a source is on this page they are using Tor.
We don't need to repeat these instructions again on another page.