Only perform the hover if the button to click is not displayed
After many strange Travis test failures, we discovered that for
whatever reason, the submit button on the route page
is not displaying properly. The hover span was being displayed
even though the initial span should be displayed.

This commit adds a check in the functional test to see if the
hover button is already displayed, so the test will go ahead
and clicks it if so without moving the cursor to the button.