Update functional tests
The functional tests were relying on identifying the ID's of the "non-hovered"
and "hovered" buttons in the old design. Changes them to test for visibility of
the specific icon image depending on the hover state.

Also makes the tests more thorough by testing the visibility of both the
`.off-hover` and `.on-hover` icon images before _and_ after hovering over the
link or button.

These tests are passing in my local development environment when I run
`./manage.py test`.