Updating "JS Warning" text
The prior text, "We recommend changing your Security Slider to high" does not identify the Tor Browser as the the requested task's location. It will be confusing to users using SecureDrop for the first time, to read this—as their brains are in SecureDrop land, and it's more intuitive at that point to look for said Security Slider in SecureDrop—not Tor. 

Also—ZOMG, Nina's first PR!! :)