Parametrizes app dev env tests
Using the `pytest.mark.parametrize` decorator to ensure that all tests
are run, even if some of the tests fail. This is a recommended best
practice with pytest, and avoids the use of `for` loops, which would
halt testing on the first failure. When testing many different pip
versions in a loop, you want to know all the versions that are wrong,
otherwise you have to patch and retest iteratively.