Converts Apache config tests Serverspec -> TestInfra
Substantial changes here. For one, the Document Interface has been
renamed to the Journalist Interface, so many tests needed logic updating
to check the correct filepath and content.

Most other checks are straightforward. We're a bit heavy handed on exact
matches for the configs, but I'm comfortable with that: configs for
these services should be well thought out, and must include updates to
the test suite, as well.