Creates conftest.py wrapper for invoking testinfra
Enables per-host variable customization via exporting to the `pytest`
namespace. Reads in YAML files based on hostname, and makes the vars
declared there available to all tests that want it.

Requires an environment variable to be exported in the testinfra wrapper
script, which is an acceptable compromise to dry out the tests.