Convert mon-staging/iptables_spec.rb -> Testinfra
Had to make some revisions to how we were doing iptables comparisons. Since
iptables is highly dependent on the order of the rules, it is not enough to
simply check iteratively whether each line exists - order matters. So now I am
comparing the entire iptables output against an expected output.