Simplifies OSSEC connectivity check
Dumbed down the test to get it to pass; the new implementation is only
applicable to local hosting via Vagrant, and will fail in CI
environments. We can patch with a more durable Ansiblized solution, as
the initial implementation strove for, once we're ready to hook up CI.

Also creates a mon-staging YAML vars file for config testing, in which
the app-staging hostname and IPv4 address are hardcoded.