Updates iptables rules for mon server
Using a simple string comparison check for the entirety of the
iptables output, which gives us solid checking about rule order. The
previous implementation supported dynamic vars, which was sharp, and
we'll definitely need to use that again. To keep moving quickly, I
marked the old test as "skipped", but did not delete it.