Fixes config test for all packages being upgraded
The test is more complicated than required because we're hardcoding an
older version of Firefox in for compatibility with Selenium. Regardless
of whether a new package will be installed, an aptitude dry-run reports:

    Would download/install/remove packages.

So we can't use that string to check whether packages would be
downloaded, installed, or removed, ironically. The new check logic
permits only Firefox as a single package to upgrade, and otherwise
all packages must be up to date (e.g. on Monitor hosts).

This check will need to be revisited once we have the prod environment
tests ported.