Sudoizes config tests where necessary
Most of the app-code tests were written targeting the development VM,
but they're relevant for the app VMs as well. Under app, grsecurity
comes into play, which forbids reading out of /proc for services not
managed by the user performing the query.

The docroots for serving up the application code are also more
restrictive on the app hosts, so let's escalate privileges there, as
well. We could technically leverage the `securedrop_user` test var
and sudo as that user to make the calls, but elevating and then
explicitly declaring exact modes will create more informative test
output in the event of failure.