Skips apt autoremove check on app-staging host
Due to the outdated version of Firefox for Selenium compatibility, we
can't assume that there will be no packages left to upgrade. Let's skip
the check only on app-staging, since on app-prod, there won't be a
Firefox or Selenium for testing, and this test doesn't run on
development since it's part of the grsecurity test suite.