Parametrizes deb package build tests
Using helper logic to keep the version definitions DRY.
Stores build directories in test YAML vars, so they're looked up
automatically, then manipulated somewhat to permit variable
interpolation within the tests. Not using proper pytest fixtures for
doing this, just basic Python. Works well enough.

Includes a commented-out line for the build path dir for
securedrop-app-code, explaining that the scripts are different for
building that package versus the others, so the shared build dir
is destroyed in between, and that makes it impossible to test the build
dir output of all. That's fine: the important thing to test the build
artifact, the .deb file itself, and that's always present post-build.