Removes redundant "unwanted" iptables config test
Since we're (rather lazily) checking for exact string equivalency of the
entire iptables rule output, we already have state=absent checking for
free; there's no need to duplicate the tests specifically for unwanted
lines.

Down the road we'll need to revisit the vars logic in the iptables
rules, since the current implementation hardcodes addresses all over the
place. Deferring work on that during test conversion; we'll tackle when
we hit the CI pipeline hard.