Partial implementation for running prod tests over Tor
The SECUREDROP_SSH_OVER_TOR=1 env var will instruct Vagrant to use the
ATHS info when constructing SSH requests. We must use the `ssh` backend
for Testinfra when running over Tor, because we don't have a dynamic
inventory file for Ansible that honors the ATHS values. So let's dump
the content of `vagrant ssh-config <hostname>` to a tempfile and use
that.