Renames "spectests" -> "config tests" in developer docs
We've ported the Serverspec tests to Testinfra, but let's generally
refer to them as "config tests" to keep things flexible and clear.
The point is that these tests validate machine state and system
configuration on the SecureDrop VMs, and the name of the particular
software project used to do so is less important.