Makes regex for HidServAuth token less strict
The tests were reporting failures on the HidServAuth token not matching
the specified regex, which was overly strict, and didn't include the `+`
character in the set for allowed chars. Let's permit any characters and
validate the length of the token, as well as its situation within the
entire HidServAuth declaration within the `hostname` file for Tor
services.