Reenables dynamic networking vars in mon-staging tests
The Jinja logic was being skipped pending a CI-compatible change, but we
need to look up the uids and gids dynamically, since they're not
consistent across builds. Fortunately @msheiny already wrote the logic,
so this change just removes the "skip" marker on the test and deletes
the YAML vars in favor of the Jinja template.