Converts iptables config tests for app-staging to dynamic vars
Using the Jinja-based templating logic, same as mon-staging, to look up
the necessary uids and gids dynamically, for interpolation in the
iptables rulesets. As described in the developer docs, these config
tests focus on testing implementation, rather than behavior, but they're
a good start toward the testing baseline we want to deliver.