Adds config test for tmux installed state
We require the tmux package to be installed, since it's used in the user
environment to provide resilience in the event that an interactive SSH
session is disconnected prematurely (as sometimes happens with Tor).

PR #1623 updates the Ansible config to ensure tmux is present, and this
test will track regressions in case it ever disappears from the base
image.