Adds /vagrant mount for libvirt in Build VM
The vagrant provisioning was falling back to rsync, which doesn't
support two-sync, so build artifacts were not present on the host
machine. The long-term solution is to use fetch for copying the source
files out and the build artifacts back, otherwise the build process
can't work in CI. Punting on that for now, though.