Copies in tests/ dir via fileglob during package build
Ensures that the tests land with the app-code package. This isn't the
best in terms of best practices, at least according to the pytest
maintainers, but it's conservative in that we've been doing it a long
while now, and it also trivially enables testing within the app-staging
environment, which otherwise would not have the tests present.