Uses rsync filters for globbing app-code files
Uses pure rsync with custom filter (i.e. include/exclude) rules in order
to prepare the contents of the build package additively. The global
exclude at the bottom of the file drops everything that wasn't
explicitly whitelisted, ensuring that we won't get .pyc files or
anything else unwanted in the future.

Also omitting the `config.py` file at the top, to make sure we don't
accidentally clobber the config.py configured via Ansible on first run.

Appending `--chown root:root` to the rsync command to ensure that the
files for the deb package have the proper ownership. In the future we
should use lintian to track checks like this.