Version-locks the sass gem in the build role
Pinning a specific version, because the "< 3.5" logic was causing
Ansible to query what's available from rubygems.org on every build. We
know the version that works well, so let's pin on that, which still
satisfies the version constraints discussed in #1594.

The associated config test was already checking the explicit version,
so removed the comment indicating that the Ansible config should be
updated to match.