Adds securedrop-keyring to the local_deb_packages list
Only applies to staging hosts, but makes sure the `securedrop-keyring`
package is present, and installs it from the locally built deb package.
The keyring package isn't at all necessary under staging, since it's
used for managing apt keyrings, and the staging machines mostly install
from local deb packages, but it's important that we keep the staging
environment config as close as possible to the config described by the
current state of the git repo, even under development.