Update inventory groups for Travis tests
For the Travis CI run, we dynamically create a simple inventory file
that targets localhost, and call a travis-specific playbook to provision
the developer environment within Travis. Long-term, we should aim to
reuse the development playbook for testing via Travis, but doing so
would require some group_vars -> host_vars conversion. In the meantime,
we'll update the inventory generation logic for Travis to match what
Vagrant creates via the group memberships.