Activates config tests for Journalist Interface logging
Tracking the Journalist Interface logging setup. Previously, due to a
misconfiguration in Apache2, the journalist interface logs showed only
the string "combined" for each and every long entry. That's no good. So
let's check that:

  1. We have log entries in that file.
  2. The log entries are never "combined".
  3. The log entries at least have the string "GET" somewhere in it.

A more aggressive testing strategy would be to validate the logformat
structure of the log entry, but that's over the top for now, and likely
to cause inaccurate test failures due to poorly thought-out regexes.