Adds app testing requirements to staging deb packages
These files were missing due to the new additive strategy for building
the app-code deb package, brought in via #1612. Without these files, one
cannot run the application tests on app-staging. Since that's definitely
something we want to do, adding those testing-specific files back into
the deb package build logic. Doing so means that the test-specific
changes will ship in prod debs, but that's OK for now, and is consistent
with what we've been doing in the past.

Note that the tests/log/firefox.log file is *not* included. This means
that everytime the staging playbook is run, the app test browser logs
will be clobbered. That's fine; the logs will still persist on
development. Including the testing log in the deb package build would
create problems for deterministic builds anyway.