Migrates Document -> Journalist interface via preinst
Implements a hook-based migration strategy for renaming the Document
Interface to Journalist Interface (for consistency with the role-based
"Source Interface" name). Services migrated include:

  * ATHS config
  * Apache vhost config

These config updates are performed as part of the `preinst` stage in the
`securedrop-app-code` deb package. Care has been taken to ensure that
the config update logic is idempotent, so that the same preinst hooks
can run on any future version of SecureDrop and ensure the same results.