Moves build package installation into build playbook
The update_version script was helpfully installing packages required for
creating a new git tag during a SecureDrop release. Since it was
implemented in bash, however, it wasn't idempotent, and it didn't have
decent informative message as to why the script was hanging for dozens
of seconds on first run (it was installing apt packages).