Softens hard requirement for CUSTOM_NOTIFICATION
A new config item was added to allow Admins to place custom strings on
the Source Interface page in #1399. The first-run Ansible config was
updated to support the new var, but the `securedrop-app-code` package
doesn't ship the `config.py` file, only `config.py.example`, so existing
production instances won't have the new value specified. The Source
Interface WSGI file should not assume the value will be defined, so
let's default to '' if it's missing. This will allow currently running
instances to upgrade cleanly, without issues.