Moves PathException back into `store.py`
The app code received a new file recently, `util.py`, that didn't have a
corresponding AppArmor profile declaration. The logic in this file has
mostly since been removed, and now includes only the custom exception
`PathException`, which previously resided in the `store.py` library.
Let's move it back there, which will neatly clean up the back-and-forth
edit history, as well as obviate the need to modify the AppArmor
profile.

Relevant commits to piece together the history are:

  * 7ae0fd9c884e4334b1b668e3af04c3ea63f5b258 # util.py created
  * 956b2137255a39fb7091a82b669af6537ba0548d # unique logic removed

The application tests check that PathException is raised when
appropriate, so we're good to go on functionality changes here.
Also removed an unused `import util` in `source.py`.