Creates app-test for custom notification setting
The config.CUSTOM_NOTIFICATION value introduced in #963 did not previously
have dedicated coverage in the app test suite. Added it. Since the app
test suite is not aware of which host it's running on (e.g. development
or app-staging), we'll check for both strings and only fail the test if
both are absent.

This should restore regression in coverage on `securedrop/source.py`,
pointed out by the failing Coveralls check in CI.