Removes force=yes in apt installs of local deb packages
The apt module doesn't support force=yes when the `deb` parameter is
used to specify the install target. It only fails on the first run;
subsequent runs are skipped with a green "OK", provided the version
string in the deb package hasn't changed. That's why we're using a
two-pass approach in the install-local-packages logic, but it would be
acceptable to use `dpkg -i` and rely on that behavior. Leaving in the
apt task until we upgrade Ansible, since recent versions may have
smarter logic around the deb parameter for the apt module.