Adds note about from address to OSSEC docs
We've had a few Admins report this problem, which is now fixed, but
poorly documented. We're leery about updating the `prod-specific.yml`
vars file with an example entry, since there's a strong risk of merge
conflict during the git stash/pop workflow, so the docs will come in
handy answering questions in the future, and enabling first-time Admins
to debug the initial setup themselves.