Adds config test for new OSSEC from address var
Off by default, so the config test is checking that the new var settings
do not override the previous default of `ossec@ossec.server` as the
sender address. If during local development the vars are changed to
support a different setup, than this will fail. I'm OK with that. While
I'd like to make the new from behavior (i.e. from address equals
authenticated account) the default, that runs a strong risk of breaking
alerts for other Admins with a working setup.