Converts Ansible mode declarations from single to double quotes
This is purely a stylistic change for the sake of consistency and
readability; it does *not* affect how Ansible processes the file modes,
so the permissions on the remote host are not altered by these changes.
If the single quotes had not been present and the double quotes were
added, that can alter the permissions intepretation, but we've been
careful about that type of regression by explicitly declaring modes in
the config test suite.