Ensures all Ansible mode declarations are quoted
A common Ansible gotcha is that file modes in multiline YAML format must
be quoted, so that they are interpreted as strings, to be passed to
`chmod`. If raw integers are provided instead, Ansible (due to Python
underneath) will interpret them as octal values and the permissions
won't map correctly, so the target host will receive (likely) broken
permissions.

To avoid such ambiguity, let's make sure that all occurances of file
mode declarations in Ansible are quoted. In the interested of
consistency, we'll use double quotes, although single quotes would have
the same effect.

Since the unquoted modes updated here already started with a leading
zero, the modes are as expected. Technically the first bit (e.g. setuid,
setgid, etc.) could have been altered, but it's sane for us to set 0 on
all these files, which are mostly config files.