Consistently uses four-digit mode declaration throughout
"400" is mostly the same as "0400" except the leading bit could
technically have been different. It is highly unlikely that bit was
every anything other than 0, but we definitely want to make sure it's 0.
It's also highly desirable to have a consistent format for mode
declarations throughout the Ansible config, which we now do.