Partial implementation of validation config test
Set to skip, so this check is currently inactive. The goal is to
validate that the playbook fails if the vars are configured incorrectly,
but part of those checks ensure the absence of staging/testing vars,
e.g. `vagrant` as username. We'll need to improve the validation logic
in a few ways prior to activating the tests:

  1. Include validation role on staging (?).
  2. Autogenerate valid prod vars on prod.
  3. Stop connecting to staging/prod hosts as `vagrant`.

We could also drop the "vagrant" check altogether and that would
substantially simplify this particular check, but we still have
hard-coded testing vars e.g. the journalist GPG key that would fail
validation.