Only assign valid tokens to Journalist.last_token
The intention of the orignal implementation of the `last_token` attribute was to
prevent a 2FA token from being re-used as it might be in some MitM or
shoulder-surfing attacks. The problem with it was that it always saved the last
token entered instead of the last valid token. Thus if attacker A watches
Journalist J use token T, they could attempt to login using T', a random token,
and then successfully login on their next attempt using T as `last_token` will
have been set to T'. This is resolved by updating `last_token` only for valid
tokens.