Consolidate tarball download tasks
Ansible is already smart enough to skip redownloading files via
`get_url` if the destination file exists. One must declare the fullpath
to the dest file, but we're already doing that. Unfortunately Ansible is
*not* smart enough to update dest files in place if the checksum doesn't
match, but that's why we have a hard fail on a subsequent task, to
handle the checksum verification.