Moves OSSEC package source trees to `install_files/`
Matches the location of all other Debian packages maintained in this
repository. Previously we had the OSSEC source packages in
`install_files/ossec-packages`, mostly as an artifact of subtreeing in
the old freedomofpress/ossec repository. Now that the OSSEC source build
logic is integrated with the SD Build VM, we it makes sense to
reorganize the directory structure so all package source trees are in a
single location.