Fix TOTP protection preventing new user login
When a user is created their `last_access` attribute is set to `None`. This
means that `totp.verify()` will default to the current time, so the protection
will always be triggered, thinking the user had just logged in. This commit adds
a check that will skip execution of the TOTP protection code if the user has
never logged in before.