Sets rsync-path to sudo for synchronize task
The `synchronize` module shipped with Ansible v1 does not intelligently
handle privilege escalation with rsync 2.6.9. In order to support both
Ansible v1 and v2 (pending resolution of #1146), and both rsync v2.6.9
(macOS) and rsync v3.1.2 (Debian Stretch), let's explictly declare a
strategy for setting the sudo privileges on the rsync task.

Full discussion of this problem can be found in comments for #1668.