Simplify TOTP MitM/shoulder-surfing protections
It's simpler to just implement a 60s wait before successful logins (not login
attempts) than to check if the token passed is one that was valid during the
last login attempt. With the protection mechanism changing as such, it made
sense to move it to `throttle_login()` and to raise a `LoginThrottledException`
instead of a `BadTokenException`. This required some corresponding changes in
`journalist.login()`.