Makes `send_encrypted_alarm.sh` resilient to GPG failure
If the GPG encryption process fails, the Monitor Server should still
send mail, but a sanitized message warning about the encryption failure.
No actual OSSEC alerts will be sent until the problem is fixed. This is
an improvement over the previous failure mode, in which OSSEC sent empty
emails if the encryption process errored out.

Also reorganizes the script into bash functions. Not only is it more
readable, it's required for using traps to notify Admins if the
encryption process fails.

The major disadvantage of the previous pipeline approach (meaning a
chain of commands joined by `|`) is that the mail would send regardless
of input or failure. Therefore the major contribution of the current
changes is the interative approach to preparing the message body,
validating that encryption worked, and only then sending.