Revert to pyotp's constant time string compare
In earlier commits, I wrongly suggest that pyotp's `compare_digest()` (used
under the hood by `strings_equal()`) constant-time string comparator was
incorrect in that it would short-circuit for every iteration through the
comparison loop after the first non-matching pair of characters were compared.
This was because I was confusing the Boolean `or` operator with the bitwise `|`
operator. The former does use short-circuit logic, but was not what we are
dealing with.

For more of my stupidity, I refer you to
https://github.com/pyotp/pyotp/pull/44#issuecomment-297901721.