Converts `fail` task in CI playbook to `assert`
The assert module is designed for this use case, checking that a list of
conditions is true, and halting execution of the playbook if any
conditions are not met. Also added a name to the task, otherwise the
playbook skips the task named "fail" if all conditions are satisfied,
which is confusing.