Synchronize securedrop-app-code dirs properly
The two-pass approach using rsync to filter files between src and
dests, both of which were remote, was unreliable: in particular, it
assumed presence of the /vagrant mounted directory, which isn't present
in CI. Instead let's synchronize the app-code directory just once.

In order to support use of the `--filter 'merge=<MERGE_FILE>'` strategy
via Ansible's synchronize module, we have to place the MERGE_FILE inside
the root of the src repo, named as `.rsync-filter`.