Adjusts build directory filepaths
The `build-generic-package.sh` script should be replaced with a more
durable solution. That's out of scope for the CI testing work, so
punting for now. Updated the bash script to use a nested approach for
storing the build directories, e.g. for `securedrop-keyring`:

  * /tmp/build/
  * /tmp/build/securedrop-keyring/
  * /tmp/build/securedrop-keyring/securedrop-keyring-0.1.0+0.3.12-amd64/
  * /tmp/build/securedrop-keyring/securedrop-keyring-0.1.0+0.3.12-amd64.deb

The two-pass approach is necessary because we don't have the version
strings exposed to the Ansible logic yet; we'll get to that in a future
change.