Updates config tests vars for sd-app-code logic
Since we're using a handful of overly specific build roles for many of
the Debian packages, we don't have a stable var structure for
determining the parent build paths. That's fine for now; the long-term
solution should be more discrete builds, e.g. within isolated
containers, leveraging a generalized build logic to ensure the packages
are consistently generated. That should set us up nicely for
deterministic builds, as well.