Moves SASS compile tasks to separate task list
Since we're no longer building from the /vagrant mounted directory, we
need to ensure that the `securedrop/sass` directory is present for
compiling the SASS to CSS. That directory is not included in the rsync
filter file, nor should it be; let's sync it in separately, compile the
SASS, then remove the directory, so it's not included in the deb
package.

Added a config test to ensure absence of all SASS-related files from the
securedrop-app-code deb package.