Removes EC2-specific logic from common reboot
The resolution of the target host address changed substantially between
Ansible v1 and v2, specifically in the context of local_action wait_for.
Seems that the use of `inventory_hostname` rather than `ansible_host` or
`ansible_ssh_host` is a sound strategy, and it seems to work well in
both CI (EC2) and local Vagrant VMs. Likely this is because both
environments use a custom SSH config file, so that SSH, not Ansible, can
translate the inventory_hostname into a routable IP.