Ensures the static/gen/ directory exists in build role
The `securedrop-app-code` deb package should include the `/gen/` static
directory, and it should be empty. Previously local builds contained
this directory, but only because the directory existed locally due to
previous invocations of the Development VM. In CI, it was revealed that
we don't actually ensure the presence of this directory anywhere; we
merely assumed it would be sychronized into the Build machine.