Ensure the .webassets-cache directory exists in build role
Also required, so that the Application Server can store a local cache of
the dynamically generated assets (via Flask-Assets). This is the result
of the additive inclusion strategy discussed in #1612: the directories
will likely exist on developer workstations (assuming the "development"
VM has been provisioned), but it won't exist in fresh clones, e.g. in
CI.

It *might* be sufficient to authorize creation of this directory in
the AppArmor profile for Apache, so the Application Server can both
create the directory and write files inside of it, in which case we can
dial down the config tests a bit. Subject to further testing.