Removes conditional on build include for staging
The conditional playbook include worked fine under Ansible v1, but under
v2 the playbook include is skipped, which means that local deb packages
are *not* rebuilt for the staging environment. That should be a hard
fail. Seems like the change was caused by the var being evaluated at
the host level, so we'd need site-wide vars to set
install_local_packages=true, which is not safe, given the repo's primary
purpose as a management tool for production SecureDrop instances.

Under the staging environment, it's safe to rebuild the packages every
time: if developers want to skip rebuilding, they can leverage tags,
e.g.:

  ANSIBLE_ARGS="--tags common" vagrant provision /staging/

which will skip the build include.