Adds sphinx pip requirements to Travis run
The sphinx libs are required for building the sphinx-based
documentation. These libs are documented in the dev docs [0],
and are being installed separate from the other pip requirements so as
not to break the config tests that check pip dependencies in the Build
VM; which, in Travis, equates to the Travis host itself.

If the build tests pass, then we lint the docs. I'd prefer it the other
way around, where a docs linting error halts execution of the test run,
but in order to achieve that we'd need better compartmentalization of
the tests.

Also includes liberal comments. Comments are good. Use comments.
Remember that you yourself, in the past, have wished for comments.
Today is your day.