Fixes "title underline too short" throughout docs (formatting)
All of these changes were related to adding asterisks to the machine
names, in accordance with the style guide, and were brought in via #1654.
Such problems generate warnings (not errors) when building the docs
locally, but in CI we're going to upgrade those warnings to errors to
keep the docs clean.