Standardizes reboot logic throughout site config
The transition of Ansible from v1 to v2 dramatically changes how the
wait_for task operates, particularly when delegated to a localhost, as
when rebooting a remote host.

Now using `ansible_ssh_host` as first value, since that's what's
hardcoded in the SecureDrop inventory files for Ansible. We'll fail over
to looking up the `inventory_hostname` as a sanity check; *not* using
the more recent `ansible_host` because that appears to map to
`localhost` when combining local_action and wait_for.