Creates developer-centric requirements file
Installs the essentials for working on SecureDrop:

  * Ansible (provisioning)
  * Testinfra (testing configuration)
  * Sphinx (documentation)

Ideally we would remove `testinfra/requirements.txt` as well, but we're
currently using that to install requirements in Travis CI as well.
Since the Travis host equates to the "development" VM for testing, we
cannot safely mess with the requirements files, as that'd cause the
config tests to fail. The proper long-term solution is better isolation
for the testing environments.