Removes mention of DigitalOcean for testing
In days of yore, we spent some time trying to wire up automated CI tests
for the staging environment in ThoughtWorks's SnapCI platform, using
DigitalOcean as a backend cloud provider for provising the staging
hosts. That work was never merged, and developers do not currently use
the `vagrant-digitalocean` plugin any longer, so it's safe to remove the
relevant config lines from the Vagrantfile.