Fixes docs ref to SSH over Tor settings
The addition of the `pip_update.sh` script to the developer docs broke
the Sphinx `ref` to SSH over Tor settings. This was missed during review
of #1674. A simple fix: simply move the ref declaration below the
pip_update.sh block, so it again correctly points to the SSH over Tor
docs.